Cleanup for service fabric
